+++ packages/sdk/src/lib/dataProtectorSharing/removeFromCollection.ts
@@ -10,13 +10,13 @@ import {
   SuccessWithTransactionHash,
 } from '../types/index.js';
 import { getSharingContract } from './smartContract/getSharingContract.js';
-import { getCollectionForProtectedData } from './smartContract/getterForSharingContract.js';
 import {
   onlyCollectionNotSubscribed,
   onlyCollectionOperator,
   onlyProtectedDataInCollection,
   onlyProtectedDataNotRented,
 } from './smartContract/preflightChecks.js';
+import { getCollectionForProtectedData } from './smartContract/sharingContract.reads.js';
 
 export const removeFromCollection = async ({
   iexec = throwIfMissing(),
