+++ packages/sdk/src/lib/dataProtectorSharing/setProtectedDataToRenting.ts
@@ -12,12 +12,12 @@ import {
   SuccessWithTransactionHash,
 } from '../types/index.js';
 import { getSharingContract } from './smartContract/getSharingContract.js';
-import { getCollectionForProtectedData } from './smartContract/getterForSharingContract.js';
 import {
   onlyCollectionOperator,
   onlyProtectedDataNotForSale,
   onlyProtectedDataInCollection,
 } from './smartContract/preflightChecks.js';
+import { getCollectionForProtectedData } from './smartContract/sharingContract.reads.js';
 
 export const setProtectedDataToRenting = async ({
   iexec = throwIfMissing(),
