+++ packages/sdk/src/lib/dataProtectorSharing/smartContract/sharingContract.reads.ts
